Create follow.cpp
This program executes a command on the command line to launch the follower package from ROS.  This program is lauched when the follower button is pressed on the tableBot gui.  With this program the user can command the robot to follow them around the room with the click of a button.